Refactor imports to use absolute paths
