+++ src/game/interfaces/ISelectable.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from game.interfaces.IHoverable import IHoverable
+from src.game.interfaces.IHoverable import IHoverable
 
 class ISelectable(IHoverable, ABC):
     pass
