+++ src/game/tiles/StoneTile.py
@@ -1,6 +1,6 @@
 import pygame
 
-from game.tiles.Tile import Tile
+from src.game.tiles.Tile import Tile
 
 
 class StoneTile(Tile):
